+++ packages/devtools_app_shared/lib/service.dart
@@ -10,6 +10,7 @@ export 'src/service/flutter_version.dart';
 export 'src/service/isolate_manager.dart' hide TestIsolateManager;
 export 'src/service/isolate_state.dart';
 export 'src/service/resolved_uri_manager.dart';
+export 'src/service/rpc_error_extension.dart';
 export 'src/service/service_extension_manager.dart'
     hide TestServiceExtensionManager;
 export 'src/service/service_manager.dart';
